Modernize code: nullptr, fixed types, safety
Large refactor across the SDK/Shared code to modernize and harden the codebase: replace NULL with nullptr, use fixed-width integer types (uint8_t, uint16_t, uint32_t, int32_t), switch pointer-sized APIs to DWORD_PTR/SDK_LPSTR where appropriate, tighten casts, and replace unsafe memory/macros with Core\macros.h. Fixed many allocations/deallocations (delete -> delete[], nulling), safer string/buffer APIs (strcpy_s/snprintf), updated resource signature strings, removed SteamCheckForWorkshop references, added EncryptedFile includes, and adjusted CMakeLists. No intented feature changes — correctness and portability improvements.

+++ Dark Basic Public Shared/Dark Basic Pro SDK/Shared/CustomBSP/CCustomBSPData.h
@@ -1,6 +1,8 @@
 #ifndef _CCUSTOMBSPDATA_H_
 #define _CCUSTOMBSPDATA_H_
 
+#include <cstdint>
+
 const int			CP_FRONT	= 1001;			// in front of plane
 const int			CP_BACK		= 1002;			// behind plane
 const int			CP_ONPLANE	= 1003;			// on plane
@@ -82,20 +84,20 @@ extern LEAF*				LeafArray;
 extern PLANE*				PlaneArray;
 extern BYTE*				PVSData;
 extern LPDIRECT3DTEXTURE9*	lpTextureSurface;
-extern WORD				NumberOfTextures;
+extern uint16_t			NumberOfTextures;
 extern POLYGON**			pTexturePolygonList;
 
-extern long				BytesPerSet;
-extern long				NumberOfPolygons;
-extern long				NumberOfNodes;
-extern long				NumberOfLeafs;
-extern long				NumberOfPlanes;
-extern long				MAXNUMBEROFNODES;
-extern long				MAXNUMBEROFPLANES;
-extern long				MAXNUMBEROFPOLYGONS;
-extern long				MAXNUMBEROFLEAFS;
+extern int32_t				BytesPerSet;
+extern int32_t				NumberOfPolygons;
+extern int32_t				NumberOfNodes;
+extern int32_t				NumberOfLeafs;
+extern int32_t				NumberOfPlanes;
+extern int32_t				MAXNUMBEROFNODES;
+extern int32_t				MAXNUMBEROFPLANES;
+extern int32_t				MAXNUMBEROFPOLYGONS;
+extern int32_t				MAXNUMBEROFLEAFS;
 
-extern long				PVSCompressedSize;
+extern int32_t				PVSCompressedSize;
 extern char				( *TextureLUT ) [ 21 ];
 
 #endif _CCUSTOMBSPDATA_H_
